CHANGE run-chromatics script to use esm
