+++ scripts/run-chromatics.js
@@ -1,18 +1,12 @@
-#!/usr/bin/env node
+import { spawn } from 'child_process';
+import { promisify } from 'util';
 
-const { spawn } = require('child_process');
-const { promisify } = require('util');
-const {
-  readdir: readdirRaw,
-  readFile: readFileRaw,
-  writeFile: writeFileRaw,
-  statSync,
-  readFileSync,
-} = require('fs');
-const { join } = require('path');
+import { readdir as readdirRaw, readFileSync } from 'fs';
+import { join } from 'path';
+
+import { getDeployables } from './utils/list-examples';
 
 const readdir = promisify(readdirRaw);
-const writeFile = promisify(writeFileRaw);
 
 const p = (l) => join(__dirname, '..', ...l);
 const logger = console;
@@ -34,20 +28,6 @@ const exec = async (command, args = [], options = {}) =>
         reject();
       });
   });
-
-const getDeployables = (files) => {
-  return files.filter((f) => {
-    const packageJsonLocation = p(['examples', f, 'package.json']);
-    let stats = null;
-    try {
-      stats = statSync(packageJsonLocation);
-    } catch (e) {
-      // the folder had no package.json, we'll ignore
-    }
-
-    return stats && stats.isFile() && hasChromaticAppCode(packageJsonLocation);
-  });
-};
 
 const hasChromaticAppCode = (l) => {
   const text = readFileSync(l, 'utf8');
@@ -100,7 +80,7 @@ const run = async () => {
   const offset = step * a;
 
   const list = examples.slice().splice(offset, step);
-  const deployables = getDeployables(list);
+  const deployables = getDeployables(list, hasChromaticAppCode);
 
   if (deployables.length) {
     logger.log(`will build: ${deployables.join(', ')}`);
